+++ src/features/people/peopleList/index.js
@@ -9,7 +9,7 @@ import { Pagination } from "../../../common/Pagination";
 import { Loader } from "../../../common/Loader";
 import { Title } from "../../../common/Title/styled";
 import { ErrorPage } from "../../../common/ErrorPage";
-import { useQueryParameter } from "../../../useQueryParameter";
+import { useQueryParameter } from "../../../queryParameters";
 import personError from "../../../images/personError.jpg";
 import {
   fetchPeople,
