--- src/useQueryParameter.js
@@ -1,8 +1,0 @@
-import { useLocation } from "react-router-dom";
-
-export const useQueryParameter = (key) => {
-    const location = useLocation();
-    const searchParams = new URLSearchParams(location.search);
-
-    return searchParams.get(key);
-};
